copy of tests for checking backwards compat
Once upon a time Tapper was named Artemis.
After heavily renaming everything to Tapper
we now need to reenable some recognition
of meta information that uses 'artemis' as
part of keys or identifiers.

That is needed to still use our huge TAP
database containing several years of test
results. Rewriting the results would be morally
more invasive than a legacy comaptibility layer.

The current tests do *not* yet test legacy
stuff but are a copy with some minor
adaptions of the current test suite.

The actual artemis stuff will be come
in the next commits.