[FEAT] Support single lazy plans without explicit section start
When we have lazy plans (i.e. plans after the of lines) we can no longer
use the plan as section start. Therefore, we used to require explicit
section start headers in that case.
For only one test plan this is in fact not necessary. Instead when we
only have one plan the we necessarily have only one section. The single
lazy plan is so common that it makes sense to handle it special. With
this change the user no longer needs to add special Tapper header to
prevent parse errors to TAP he gets out of standard TAP tools.

cr https://cr.amazon.com/r/1536480/