[FIX] section names that lead to illegal file names for generate_html
Some section names lead to file names that are somehow illegal. One
example for this is a section name that contains just a dot. This leads
to the creation of a file for html generation that has the name
"section/.". Also, any section name containing a slash (/) would lead to
an illegal file name. This patch creates a file name for reports with
such section names that have dots and slashed substituted with the word
dot or slash respectively.